🐛 Bug Fix!
Sometimes when you attempt to sync a large amount of components at once, some of them gets ignored and won't be affected.
This is due the fact we're using a loop to call render() on each one of them, and this could go so fast the way that react cannot handle all of these updates.
A good solution would be wrapping render() inside a timeout just like that: setTimeout(() => render(), 0).